Resolve content-type of CallbackEvent attachments on save
Before we can render the correct template for the correct content
type, we need to know what content type the given attachment is.

This commit allows to resolve the content type of the attachment
on save by first checking if the file extension is of a known
filetype, and if not - looking up the content type on the server
with a simple HEAD request.

Test coverage increased to cover saving new CallbackEvent models
and new logic.