Add attachment event template + rendering filters
The test_app now has the ability to act on addAttachmentToCard
trello events and render the template including the right
render if it has an image attachment.

Currently the rendering code is a simple branch if its an image; if
more support for new content types comes along - a more thorough
approach will be needed (likely a dictionary switch with seperate
render functions per content type).